+++ src/main/kotlin/com/github/gib/services/GivServiceSettings.kt
@@ -5,19 +5,21 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
+import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.util.Pair
 import com.intellij.util.ui.JBImageIcon
 import java.net.URL
 import javax.imageio.ImageIO
 
-@State(name = "GivServiceSettings", storages = [Storage("editor.xml"), Storage("workspace.xml", deprecated = true)])
+@State(name = "GivServiceSettings", storages = [Storage("gideabrowser.xml")])
 class GivServiceSettings : PersistentStateComponent<GivServiceSettings.State> {
 
     private val lock = Any()
     private var myState = State()
 
-
     companion object {
+        private val LOG = logger<GivServiceSettings>()
+
         @JvmStatic
         fun instance(): GivServiceSettings {
             return ApplicationManager.getApplication().getService(GivServiceSettings::class.java)
@@ -45,15 +47,32 @@ class GivServiceSettings : PersistentStateComponent<GivServiceSettings.State> {
     fun addToFavorites(webToFavorite: List<String>) = synchronized(lock) {
         myState.favorites.removeAll { true }
         webToFavorite.forEach {
-            myState.favorites.add(Pair.create(it,
-                JBImageIcon(ImageIO.read(URL("https://www.google.com/s2/favicons?domain=$it")))))
+            try {
+                val imageIo = JBImageIcon(ImageIO.read(URL("https://www.google.com/s2/favicons?domain=$it")))
+                myState.favorites.add(
+                    Pair.create(
+                        it,
+                        imageIo
+                    )
+                )
+            } catch (e: Exception) {
+                LOG.warn("Error adding favorite $it", e)
+            }
         }
     }
 
     fun addFavorite(webToFavorite: String) = synchronized(lock) {
-        myState.favorites.add(Pair.create(webToFavorite,
-            JBImageIcon(ImageIO.read(URL("https://www.google.com/s2/favicons?domain=$webToFavorite")))))
-
+        try {
+            val imageIo = ImageIO.read(URL("https://www.google.com/s2/favicons?domain=$webToFavorite"))
+            myState.favorites.add(
+                Pair.create(
+                    webToFavorite,
+                    JBImageIcon(imageIo)
+                )
+            )
+        } catch (e: Exception) {
+            LOG.warn("Error adding favorite", e)
+        }
     }
 
     fun addToHeadersOverwrite(headersOverwrite: List<HeadersOverwrite>) = synchronized(lock) {
@@ -61,9 +80,9 @@ class GivServiceSettings : PersistentStateComponent<GivServiceSettings.State> {
     }
 
     class State {
-        var homePage: String = "https://plugins.jetbrains.com/plugin/18269-queryflag"
+        var homePage: String = "https://www.dorkag.com/dorkag"
         val favorites = mutableListOf<Pair<String, JBImageIcon>>()
-        var headersOverwrite = List<HeadersOverwrite>(0) { HeadersOverwrite("", "","",false) }
+        var headersOverwrite = List(0) { HeadersOverwrite("", "", "", false) }
     }
 
     override fun getState(): State = synchronized(lock) {
@@ -73,4 +92,5 @@ class GivServiceSettings : PersistentStateComponent<GivServiceSettings.State> {
     override fun loadState(state: State) = synchronized(lock) {
         myState = state
     }
+
 }